Moved CHR$
